+++ src/core.rs
@@ -1,12 +1,18 @@
+pub(crate) use self::docker::Docker;
 pub use self::{
     container::Container,
-    docker::{Docker, Logs, Ports, RunArgs},
+    container_async::ContainerAsync,
+    docker::RunArgs,
     image::{Image, Port, WaitFor},
-    wait_for_message::{WaitError, WaitForMessage},
 };
 
 mod container;
+mod container_async;
 mod docker;
 pub mod env;
 mod image;
-mod wait_for_message;
+
+pub(crate) mod logs;
+pub(crate) mod ports;
+
+pub(crate) use container_async::DockerAsync;