+++ src/core/docker.rs
@@ -1,19 +1,4 @@
-use crate::{core::Port, Container, Image};
-use std::{collections::HashMap, fmt, io::Read};
-
-/// Defines the minimum API required for interacting with the Docker daemon.
-pub trait Docker
-where
-    Self: Sized,
-{
-    fn run<I: Image>(&self, image: I) -> Container<'_, Self, I>;
-    fn run_with_args<I: Image>(&self, image: I, run_args: RunArgs) -> Container<'_, Self, I>;
-    fn logs(&self, id: &str) -> Logs;
-    fn ports(&self, id: &str) -> Ports;
-    fn rm(&self, id: &str);
-    fn stop(&self, id: &str);
-    fn start(&self, id: &str);
-}
+use crate::core::{logs::LogStream, ports::Ports, Port};
 
 /// Container run command arguments.
 /// `name` - run image instance with the given name (should be explicitly set to be seen by other containers created in the same docker network).
@@ -24,6 +9,19 @@ pub struct RunArgs {
     name: Option<String>,
     network: Option<String>,
     ports: Option<Vec<Port>>,
+}
+
+/// Defines operations that we need to perform on docker containers and other entities.
+///
+/// This trait is pub(crate) because it should not be used directly by users but only represents an internal abstraction that allows containers to be generic over the client they have been started with.
+/// All functionality of this trait is available on [`Container`]s directly.
+pub(crate) trait Docker {
+    fn stdout_logs(&self, id: &str) -> LogStream;
+    fn stderr_logs(&self, id: &str) -> LogStream;
+    fn ports(&self, id: &str) -> Ports;
+    fn rm(&self, id: &str);
+    fn stop(&self, id: &str);
+    fn start(&self, id: &str);
 }
 
 impl RunArgs {
@@ -58,39 +56,5 @@ impl RunArgs {
 
     pub(crate) fn ports(&self) -> Option<Vec<Port>> {
         self.ports.clone()
-    }
-}
-
-/// The exposed ports of a running container.
-#[derive(Debug, PartialEq, Default)]
-pub struct Ports {
-    mapping: HashMap<u16, u16>,
-}
-
-impl Ports {
-    /// Registers the mapping of an exposed port.
-    pub fn add_mapping(&mut self, internal: u16, host: u16) -> &mut Self {
-        log::debug!("Registering port mapping: {} -> {}", internal, host);
-
-        self.mapping.insert(internal, host);
-
-        self
-    }
-
-    /// Returns the host port for the given internal port.
-    pub fn map_to_host_port(&self, internal_port: u16) -> Option<u16> {
-        self.mapping.get(&internal_port).cloned()
-    }
-}
-
-/// Log streams of running container (stdout & stderr).
-pub struct Logs {
-    pub stdout: Box<dyn Read>,
-    pub stderr: Box<dyn Read>,
-}
-
-impl fmt::Debug for Logs {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("Logs").finish()
     }
 }